backend/requirements.txt
# Web Server & Framework
fastapi==0.128.0
uvicorn==0.34.0
python-multipart==0.0.20

# Data Validation & Settings
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.1

# AI & Orchestration
langchain==0.3.14
langchain-openai==0.2.14
langchain-google-genai==4.2.0
langgraph==0.2.60

# Document Processing
pymupdf==1.25.1